+++ pyscenarios/__init__.py
@@ -1,22 +1,19 @@
-import pkg_resources
+import importlib.metadata
 
-from .copula import gaussian_copula, t_copula
-from .sobol import max_sobol_dimensions, sobol
-from .stats import tail_dependence
+from pyscenarios.copula import gaussian_copula, t_copula
+from pyscenarios.sobol import max_sobol_dimensions, sobol
+from pyscenarios.stats import tail_dependence
 
 try:
-    __version__ = pkg_resources.get_distribution("pyscenarios").version
-except Exception:  # pragma: nocover
-    # Local copy or not installed with setuptools.
-    # Disable minimum version checks on downstream libraries.
+    __version__ = importlib.metadata.version("pyscenarios")
+except importlib.metadata.PackageNotFoundError:  # pragma: nocover
+    # Local copy, not installed with pip
     __version__ = "999"
 
 # Prevent Intersphinx from pointing to the implementation modules
 for obj in gaussian_copula, t_copula, sobol, max_sobol_dimensions, tail_dependence:
     obj.__module__ = "pyscenarios"
-
 del obj
-del pkg_resources
 
 __all__ = (
     "__version__",