+++ pyscenarios/stats.py
@@ -1,14 +1,16 @@
 """Statistical functions
 """
-from typing import Any, Union
+from __future__ import annotations
+
+from typing import Any
 
 import dask.array as da
 import numpy as np
 
-from . import duck
+from pyscenarios import duck
 
 
-def tail_dependence(x: Any, y: Any, q: Any) -> Union[np.ndarray, da.Array]:
+def tail_dependence(x: Any, y: Any, q: Any) -> np.ndarray | da.Array:
     r"""Calculate `tail dependence
     <https://en.wikipedia.org/wiki/Tail_dependence>`_
     between vectors x and y.